PLANNER-252 Move VariableListenerSupport into subpackage support
(cherry picked from commit 3d88aed4ab5fa3bec19df32a005c3558812f34eb)
